Moves code needed to create fixtures to fixtureable module
